+++ hertzbeat-alerter/src/main/java/org/apache/hertzbeat/alert/dto/TencentCloudExternAlert.java
@@ -1,0 +1,128 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.hertzbeat.alert.dto;
+
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ *
+ * Tencent cloud alarm entity class
+ */
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class TencentCloudExternAlert {
+    @JsonProperty("sessionID")
+    private String sessionId;
+    private String alarmStatus;
+    private String alarmType;
+    private AlarmObjInfo alarmObjInfo;
+    private AlarmPolicyInfo alarmPolicyInfo;
+    private String firstOccurTime;
+    private int durationTime;
+    private String recoverTime;
+
+    /**
+     * Alarm Object information
+     */
+    @Data
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class AlarmObjInfo {
+        private String region;
+        private String namespace;
+        @JsonProperty("appID")
+        private String appId;
+        private String uin;
+        private Dimensions dimensions;
+    }
+
+    /**
+     *  Uniform Resource ID information
+     */
+    @Data
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class Dimensions {
+        @JsonProperty("unInstanceID")
+        private String unInstanceId;
+        @JsonProperty("objID")
+        private String objId;
+    }
+
+    /**
+     * Alarm policy information
+     */
+    @Data
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class AlarmPolicyInfo {
+        @JsonProperty("policyID")
+        private String policyId;
+        private String policyType;
+        private String policyName;
+        @JsonProperty("policyTypeCName")
+        private String policyTypeCname;
+        private Conditions conditions;
+    }
+
+    /**
+     * Parameters of indicator alarms
+     */
+    @Data
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class Conditions {
+        // alarm metrics parameters
+        private String metricName;
+        private String metricShowName;
+        private String calcType;
+        private String calcValue;
+        private String calcUnit;
+        private String currentValue;
+        private String historyValue;
+        private String unit;
+        private String period;
+        private String periodNum;
+        private String alarmNotifyType;
+        private long alarmNotifyPeriod;
+
+        // alarm event parameters
+        private String productName;
+        private String productShowName;
+        private String eventName;
+        private String eventShowName;
+    }
+
+    /**
+     * Transaction alarm
+     */
+    private static final String EVENT = "event";
+
+    /**
+     * Indicator alarm
+     */
+    private static final String METRIC = "metric";
+
+}
