+++ hertzbeat-alerter/src/main/java/org/apache/hertzbeat/alert/service/ExternAlertService.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.hertzbeat.alert.service;
+
+/**
+ * Extern alert service for prometheus, cloud and so on
+ */
+public interface ExternAlertService {
+
+    /**
+     * Save external alarms of services
+     * @param content alert report json string
+     */
+    void addExternAlert(String content);
+
+
+    /**
+     * Support source: prometheus, tencent ...
+     * @return source name
+     */
+    String supportSource();
+}
