+++ web-app/src/assets/doc/alert-integration/prometheus.en-US.md
@@ -1,0 +1,7 @@
+# Prometheus Alert Integration
+
+HertzBeat is fully compatible with Prometheus alert data format. You can configure Prometheus alerting rules to send alerts to HertzBeat.
+
+## Prometheus Alert Configuration
+
+Add the following configuration to your Prometheus configuration file:
