+++ web-app/src/assets/doc/alert-integration/tencent.zh-TW.md
@@ -1,0 +1,9 @@
+## 验证配置
+
+1. 确保 Prometheus 配置正确并重新加载配置
+2. 触发一个测试告警
+3. 在 HertzBeat 告警中心查看是否收到告警
+
+更多信息请参考 [Prometheus 官方文档](https://prometheus.io/docs/alerting/latest/configuration/)
+
+## Alert Rule Example
